refactor: Removing legacy code and restructring
This change removes some of the legacy code around contract deployment
that's not being used and very out of date by now. It also extracts all
smart contract logic into a seperate file to add a bit seperation of
concern.

Another addressed topic here is the differentiation between the compass
EVM contract and arbitrary contracts deployed by validators without
undergoing governance.

The change also includes scaffolding for unilateral contract deployment.
