--- x/consensus/client/cli/tx_add_messages_signatures.go
@@ -1,39 +1,0 @@
-package cli
-
-import (
-	"strconv"
-
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/spf13/cobra"
-)
-
-var _ = strconv.Itoa(0)
-
-func CmdAddMessagesSignatures() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "add-messages-signatures",
-		Short: "Broadcast message AddMessagesSignatures",
-		Args:  cobra.ExactArgs(0),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			// TODO: we don't need to have this from the CLI
-			panic("don't use from cli")
-
-			//clientCtx, err := client.GetClientTxContext(cmd)
-			//if err != nil {
-			//	return err
-			//}
-
-			//msg := types.NewMsgAddMessagesSignatures(
-			//	clientCtx.GetFromAddress().String(),
-			//)
-			//if err := msg.ValidateBasic(); err != nil {
-			//	return err
-			//}
-			//return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
-		},
-	}
-
-	flags.AddTxFlagsToCmd(cmd)
-
-	return cmd
-}
