+++ x/evm/client/cli/tx.go
@@ -21,5 +21,7 @@ func GetTxCmd() *cobra.Command {
 		RunE:                       client.ValidateCmd,
 	}
 
+	cmd.AddCommand(CmdDeploySmartContract())
+
 	return cmd
 }
