+++ x/evm/keeper/attest_test.go
@@ -174,6 +174,9 @@ var _ = g.Describe("attest router", func() {
 				sc, err := k.SaveNewSmartContract(ctx, contractAbi, common.FromHex(contractBytecodeStr))
 				Expect(err).To(BeNil())
 
+				err = k.SetAsCompassContract(ctx, sc)
+				Expect(err).To(BeNil())
+
 				dep, _ := k.getSmartContractDeploying(ctx, sc.GetId(), newChain.GetChainReferenceID())
 				Expect(dep).NotTo(BeNil())
 			}
