--- x/evm/keeper/msg_server_submit_new_job.go
@@ -1,12 +1,0 @@
-package keeper
-
-import (
-	"context"
-
-	"github.com/palomachain/paloma/x/evm/types"
-)
-
-func (k msgServer) SubmitNewJob(goCtx context.Context, msg *types.MsgSubmitNewJob) (*types.MsgSubmitNewJobResponse, error) {
-	// TODO: remove this
-	panic("do not use this")
-}
