+++ x/evm/module.go
@@ -158,7 +158,7 @@ func (am AppModule) BeginBlock(_ sdk.Context, _ abci.RequestBeginBlock) {}
 // EndBlock executes all ABCI EndBlock logic respective to the capability module. It
 // returns no validator updates.
 func (am AppModule) EndBlock(ctx sdk.Context, _ abci.RequestEndBlock) []abci.ValidatorUpdate {
-	am.keeper.TryDeployingLastSmartContractToAllChains(ctx)
+	am.keeper.TryDeployingLastCompassContractToAllChains(ctx)
 
 	if ctx.BlockHeight()%300 == 0 {
 		func() {
