--- x/evm/simulation/submit_new_job.go
@@ -1,29 +1,0 @@
-package simulation
-
-import (
-	"math/rand"
-
-	"github.com/cosmos/cosmos-sdk/baseapp"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
-	"github.com/palomachain/paloma/x/evm/keeper"
-	"github.com/palomachain/paloma/x/evm/types"
-)
-
-func SimulateMsgSubmitNewJob(
-	ak types.AccountKeeper,
-	bk types.BankKeeper,
-	k keeper.Keeper,
-) simtypes.Operation {
-	return func(r *rand.Rand, app *baseapp.BaseApp, ctx sdk.Context, accs []simtypes.Account, chainReferenceID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-		simAccount, _ := simtypes.RandomAcc(r, accs)
-		msg := &types.MsgSubmitNewJob{
-			Creator: simAccount.Address.String(),
-		}
-
-		// TODO: Handling the SubmitNewJob simulation
-
-		return simtypes.NoOpMsg(types.ModuleName, msg.Type(), "SubmitNewJob simulation not implemented"), nil, nil
-	}
-}
